+++ packages/metalcore/src/metalcore/householder.py
@@ -1,7 +1,8 @@
-# Author: Kris Bailey
-# Copyright 2026
-# Email: kris@krisbailey.com
 """
+Author: Kris Bailey
+Copyright 2026
+Email: kris@krisbailey.com
+
 Householder Reflections - Core orthogonalization primitive.
 
 A Householder reflection is an orthogonal transformation: