+++ packages/metalcore/src/metalcore/solve.py
@@ -1,7 +1,8 @@
-# Author: Kris Bailey
-# Copyright 2026
-# Email: kris@krisbailey.com
 """
+Author: Kris Bailey
+Copyright 2026
+Email: kris@krisbailey.com
+
 Linear System Solver using QR Factorization
 
 Solves Ax = b for square systems using QR decomposition: