+++ packages/metalcore/src/metalcore/trsm.py
@@ -1,7 +1,8 @@
-# Author: Kris Bailey
-# Copyright 2026
-# Email: kris@krisbailey.com
 """
+Author: Kris Bailey
+Copyright 2026
+Email: kris@krisbailey.com
+
 Triangular Solve (trsm) - Solve Ax = b where A is triangular.
 
 This is a foundational operation used by: